refactor(android): centralize recurrence calculation in RecurrenceUtils
Extract RRULE next-occurrence parsing/calculation into a shared utility and reuse it in CalendarWidget and NotificationWorker to remove duplication and keep recurrence behavior consistent.
